requirements.txt
chainlit>=1.0.0
azure-ai-projects>=1.0.0b12
azure-identity>=1.15.0
python-dotenv>=1.0.0
aiohttp>=3.8.1
